requirements.txt
deepgram-sdk==3.7.4
ffmpeg==1.4
groq==0.9.0
numpy==1.26.4
openai==1.43.1
pandas==2.2.2
pydub==0.25.1
python-dotenv==1.0.0
requests==2.32.3
scikit-learn==1.5.1
sounddevice==0.5.1
soundfile==0.12.1
websocket-client==1.8.0
websockets==13.1
PyAudio==0.2.14